Update README.md
- 1. Task Description
The team is tasked with designing a Voting System following Scrum methodologies. Each team member is assigned specific roles:

Team Leader (SCRUM Master): Responsible for guiding and coaching the Scrum team, removing impediments, and ensuring adherence to Scrum practices.
Secretary: Assists the team leader with administrative tasks, focusing on organizational and administrative support.
2x Developers: A cross-functional team responsible for delivering a potentially shippable product increment at the end of every Sprint. They collaborate closely and share accountability for the success of each Sprint.
1x Tester: Responsible for ensuring the quality and reliability of the software by identifying and fixing defects or issues.
The team must select and implement an appropriate data structure to store data details for searching. Time complexity analysis of the key operations using this data structure is also required.
- 2. Scenario
The team's program must use an appropriate data structure to store voting data and allow the user to specify the input file name. The focus is on the library and search functionality.